Change wifi configuration setting
4 values are passed from the cam menu, WIFI_MODE, AP_COUNTRY,
WIFI_MAC and AP_CHANNEL_5G. These paramaters are used to create
a default wifi.conf.
If a wifi.conf file exists on the SDCard, it is used otherwise
the default one is used.
None of the 4 passed values need to be edited in the wifi.conf
file on the SDCard unless the user want to force them.