Fix station mode when no AP
This fix the station mode when the needed AP is not present.
Send a notification that the cam is connected to a AP
named FAKE.
The wifi can not be switch off by this script otherwise
this freeze the cam on the wifi switch on/off menu.
So manual switch off have to be done.