Add a conversion method for files on the SDCard
Following an issue https://dashcamtalk.com/forum/threads/yi-4k-z16v13l-custom-firmware-and-more-update-18-08-08.35439/post-424998
If files are not with the right EOL, they are useless.
Use dos2unix to convert EOL on all files edited by users.
By the way, files have to be copied to tmp folder, no modifications
of files are able on the SDCard.